Catch IllegalArgumentException with primitives
It was possible to pass null to a method expecting a primitive
parameter type, resulting in an unhandled IllegalArgumentException.

We now catch that, and other throwables, and present them the same as
anything else.

I debated catching this and ignoring it, sending the method some default
value, however I don't think we should be hiding that from the user. The
point of this plugin is show how the server responds, as such I see no
reason to suppress that behavior.