Make PlatformUtil check for the BaseComponent API
This will allow it to be run only in the presence of Spigot-API or ...
some other API that implements bungee's chat API. Prior to this we were
checking for Spigot's implementation classes which doesn't make much
sense seeing as we never use them anyway.

Also makes the console log the inner exception, if present, when it runs
into an exception. Generally, these are what we're most interested in.